# replit-app/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PlayerColor = typeof PlayerColor[keyof typeof PlayerColor];


export const PlayerColor = {
  red: 'red',
  green: 'green',
  yellow: 'yellow',
  blue: 'blue',
} as const;

export interface Player {
  id: string;
  name: string;
  /** @nullable */
  username?: string | null;
  color: PlayerColor;
  isHost: boolean;
  isSpectator: boolean;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  waiting: 'waiting',
  playing: 'playing',
  finished: 'finished',
} as const;

export interface Room {
  roomCode: string;
  status: RoomStatus;
  players: Player[];
  spectators: number;
  currentPlayer: number;
  /** @nullable */
  dice: number | null;
  /** @nullable */
  winner: number | null;
  tokens: number[][];
}

export type RoomInputMode = typeof RoomInputMode[keyof typeof RoomInputMode];


export const RoomInputMode = {
  classic: 'classic',
  quick: 'quick',
} as const;

export interface PlayerInput {
  /** @minLength 1 */
  id: string;
  /**
     * @minLength 1
     * @maxLength 64
     */
  name: string;
  /** @nullable */
  username?: string | null;
}

export interface RoomInput {
  player: PlayerInput;
  mode?: RoomInputMode;
}

export interface ErrorResponse {
  error: string;
}