# replit-app/lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerColor } from './playerColor';

export interface Player {
  id: string;
  name: string;
  /** @nullable */
  username?: string | null;
  color: PlayerColor;
  isHost: boolean;
  isSpectator: boolean;
}