# replit-app/lib/api-zod/src/generated/types/playerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerInput {
  /** @minLength 1 */
  id: string;
  /**
     * @minLength 1
     * @maxLength 64
     */
  name: string;
  /** @nullable */
  username?: string | null;
}