# replit-app/lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';
import type { RoomStatus } from './roomStatus';

export interface Room {
  roomCode: string;
  status: RoomStatus;
  players: Player[];
  spectators: number;
  currentPlayer: number;
  /** @nullable */
  dice: number | null;
  /** @nullable */
  winner: number | null;
  tokens: number[][];
}