# replit-app/lib/api-zod/src/generated/types/roomInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerInput } from './playerInput';
import type { RoomInputMode } from './roomInputMode';

export interface RoomInput {
  player: PlayerInput;
  mode?: RoomInputMode;
}